Title: 합과 평균을 구하는 함수, Time: 103ms, Memory: 8MB, Status: Passed - CodeTree

+++ 240209/합과 평균을 구하는 함수/function-that-obtains-sum-and-mean.java
@@ -14,6 +14,6 @@ public class Main {
         }
 
         System.out.println(sum);
-        System.out.println(Math.round(sum/3));
+        System.out.println((int)Math.round((double)sum/3));
     }
 }
